requirements.txt
# ==== Core Unstructured (PDF, imagens, DOCX, EML, etc.) ====
# unstructured[all-docs]==0.18.16
unstructured-inference>=0.7,<0.8
pytesseract>=0.3.10
pdf2image>=1.16.3
pypdf>=4.2
pdfminer.six>=20231228
Pillow>=10.3
opencv-python>=4.8
onnxruntime>=1.17
unstructured[all-docs]
unstructured[openai]

# ==== Utilidades e parsers auxiliares (recomendado) ====
python-magic-bin
filetype>=1.2
beautifulsoup4>=4.12
lxml>=5.2
markdown-it-py>=3.0
python-docx>=1.1
openpyxl>=3.1
msg-parser>=1.2
chardet>=5.2
tqdm>=4.66

# ==== Stack de dados / notebook (para o .ipynb) ====
jupyterlab>=4.2
ipykernel>=6.29
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.8
seaborn>=0.13

# ==== Integração com Neo4j e OpenAI ====
python-dotenv 
neo4j 
openai
tiktoken>=0.7.0
ujson>=5.10.0
